fix bug: maxcompressLen calc error, UINT32 will overflow
git-svn-id: http://192.168.20.11/sequoiadb/trunk@29778 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/util/utilCompressorLZW.cpp
@@ -449,7 +449,7 @@ namespace engine
          level = strategy->_level ;
       }
 
-      maxSize = sourceLen * minRatio / 100 ;
+      maxSize = (UINT32)( (UINT64)sourceLen * minRatio / 100 ) ;
 
       dictionary.attach( (void *)dictHandle ) ;
       context.setDictionary( &dictionary ) ;
